# Speed up regular result iterator