--- a/src/manage_sql.c
+++ b/src/manage_sql.c
@@ -22072,6 +22072,9 @@ init_result_get_iterator (iterator_t* iterator, const get_data_t *get,
   else
     actual_columns = columns_no_cert;
 
+/**
+ * @brief SQL for getting current severity.
+ */
 #define CURRENT_SEVERITY_SQL                                            \
   "coalesce ((CASE WHEN results.severity > " G_STRINGIFY (SEVERITY_LOG) \
   "           THEN CAST (nvts.cvss_base AS double precision)"           \