**What**:

Speed up iterator created by init_result_get_iterator:

1. Simplify level check when checking for all levels.
2. Use LATERAL to keep the new severity SQL in a single place in the statement.
3. Inline SQL from views and functions to take advantage of the JOIN on table nvts.
4. Use WITH to isolate the current severity calculation in the "overrides on, dynamic on" case.

Improvements to my GET_RESULTS times:
````
static, overrides off      17s to 6s
static, overrides on      248s to 34s
dynamic, overrides off     45s to 6s
dynamic, overrides on     415s to 33s
````
This translates to similar jumps in the GSA Results page.

This PR is similar to the changes to the result counting iterator in https://github.com/greenbone/gvmd/pull/1358.

<!--
  Describe what changes are being made, e.g. which feature/bug is being
  developed/fixed in this PR?
-->

**Why**:

Improves the speed of the GSA result pages when using bigger dbs, especially the Results page.

Making the dynamic cases fast may enable us to switch to "query-time" dynamic-only severity (https://github.com/greenbone/gvmd/pull/1337).

<!-- Why are these changes necessary? -->

**How did you test it**:

<details>
  <summary>With an 800k result db, having 3 overrides.</summary><p>

 1. `git checkout master`
 2. `make install`
 3. `gvmd`
 4. `update settings set value = 0 where owner = (select id from users where name = 'm') and uuid = '77ec2444-e7f2-4a80-a59b-f4237782d93f';`
 5. `time o m m '<get_results details="1" filter="apply_overrides=0 min_qod=70 sort-reverse=severity rows=10 first=1"/>' > /tmp/master-st` => 17s
 6. `time o m m '<get_results details="1" filter="apply_overrides=1 min_qod=70 sort-reverse=severity rows=10 first=1"/>' > /tmp/master-st-ov` => 248s
 7. `update settings set value = 1 where owner = (select id from users where name = 'm') and uuid = '77ec2444-e7f2-4a80-a59b-f4237782d93f';`
 8. `time o m m '<get_results details="1" filter="apply_overrides=0 min_qod=70 sort-reverse=severity rows=10 first=1"/>' > /tmp/master-dy` => 45s
 9. `time o m m '<get_results details="1" filter="apply_overrides=1 min_qod=70 sort-reverse=severity rows=10 first=1"/>' > /tmp/master-dy-ov` => 415s
 9. `time o m m '<get_results details="1" filter="rexec apply_overrides=1 min_qod=70 sort-reverse=severity rows=10 first=1"/>' > /tmp/master-dy-ov-filter` => 209s (with pg jit off)

 10. `git checkout lateral-results`
 11. `make install`
 12. `gvmd`
 13. `update settings set value = 1 where owner = (select id from users where name = 'm') and uuid = '77ec2444-e7f2-4a80-a59b-f4237782d93f';`
 14. `time o m m '<get_results details="1" filter="apply_overrides=0 min_qod=70 sort-reverse=severity rows=10 first=1"/>' > /tmp/lateral-st` => 6s
 15. `time o m m '<get_results details="1" filter="apply_overrides=1 min_qod=70 sort-reverse=severity rows=10 first=1"/>' > /tmp/lateral-st-ov` => 34s
 16. `update settings set value = 1 where owner = (select id from users where name = 'm') and uuid = '77ec2444-e7f2-4a80-a59b-f4237782d93f';`
 17. `time o m m '<get_results details="1" filter="apply_overrides=0 min_qod=70 sort-reverse=severity rows=10 first=1"/>' > /tmp/lateral-dy` => 5s
 18. `time o m m '<get_results details="1" filter="apply_overrides=1 min_qod=70 sort-reverse=severity rows=10 first=1"/>' > /tmp/lateral-dy-ov` => 33s
 18. `time o m m '<get_results details="1" filter="rexec apply_overrides=1 min_qod=70 sort-reverse=severity rows=10 first=1"/>' > /tmp/lateral-dy-ov-filter` => 73s (with pg jit off)


 19. `cd /tmp`
 20. `diff master-st lateral-st` => empty
 21. `diff master-st-ov lateral-st-ov`  => empty
 22. `diff master-dy lateral-dy` => empty
 23. `diff master-dy-ov lateral-dy-ov` => empty
24. `diff master-dy-ov-filter lateral-dy-ov-filter` => empty
</p>
</details>

<!--
  How did you verify the changes in this PR?
  If this PR contains tests this section can be considered done.
  Otherwise please write down the steps on how you did test your changes and
  verified that the changes are working as expected.

  See https://www.ministryoftesting.com/dojo/lessons/community-stories-to-shift-left-start-right
  for some background.
 -->

**Checklist**:

<!-- add "N/A" to the end of each line not applicable to your changes -->

<!-- to check an item, place an "x" in the box like so: "- [x] Tests" -->

- [ ] Tests
- [x] [CHANGELOG](https://github.com/greenbone/gvmd/blob/master/CHANGELOG.md) Entry